% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/europe_map_creation.R
\name{europe_map_creation}
\alias{europe_map_creation}
\title{European Maps at the NUTS-2 Level}
\usage{
europe_map_creation(
  map_data,
  type = "ALL",
  title_t = "",
  subtitle_t = "",
  caption_t = "",
  var_name = "Legend",
  col_pal = "ag_GrnYl"
)
}
\arguments{
\item{map_data}{Data frame with two columns. The first should have the NUTS-2
codes for the regions to be visualised, the second should have the values for
each region}

\item{type}{A variable that takes the values "CONTINENTAL" to visualise only
continental europe, "ALL" to visualise all territories including extraregions,
and a NUTS-0 country code to visualise a map of that country}

\item{title_t}{Text to be displayed as the title of the graph}

\item{subtitle_t}{Text to be displayed as the subtitle of the graph}

\item{caption_t}{Text to be displayed as the caption of the graph}

\item{var_name}{Text to be displayed as the title of the legend}

\item{col_pal}{Name of the palette to be used when making the graph, acording
to the HCL colour palettes available in the grDevices package}
}
\description{
Create European maps at the NUTS-2 level. Adjust to visualise only continental
europe, adding the extra regions, or only one country.
}
