% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_map_creation.R
\name{region_map_creation}
\alias{region_map_creation}
\title{Regional Map Creation at the NUTS-2 level}
\usage{
region_map_creation(
  map_data,
  type = "NAT",
  global_map,
  pal,
  labs_plot,
  var_name
)
}
\arguments{
\item{map_data}{Data from which the map should be created.}

\item{type}{Variable that can take the values of "NAT" if it is the National
or continental level or the name of the extra region to create such as
"Martinique" or "Canary Islands".}

\item{global_map}{Optional parameter, only used when type = "NAT", it is used
as the background mute global map}

\item{labs_plot}{Vector of labels to be used in the scale of the map}

\item{var_name}{Vector of values to be used as the scale of the map}
}
\description{
Create a map for a specific region. Helper function for the europe_map_creation()
function inside the EUspiR package.
}
\keyword{internal}
