% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi_normalisation.R
\name{spi_normalisation}
\alias{spi_normalisation}
\title{Normalisation function within EU-SPI (2020)}
\usage{
spi_normalisation(val, dis_val, uto_val, inverted)
}
\arguments{
\item{val}{Value that needs to be normalised}

\item{dis_val}{Dystopic value of the indicator}

\item{uto_val}{Utopic value of the indicator}

\item{inverted}{Logical value. Does the indicator have a positive relation
with the index (inverted = FALSE) or a negative relation with the index
(inverted = TRUE)}
}
\value{
Normalised numeric value.
}
\description{
Normalise the data according to the methodology seen in the EU-SPI (2020). It
only works with single values; for vectors or matrices, see \link{spi_normalisation_m}.
}
\examples{
spi_normalisation(20,0,100,TRUE)
spi_normalisation(10,100,0,FALSE)

}
