% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi_normalisation_data.R
\docType{data}
\name{spi_normalisation_data}
\alias{spi_normalisation_data}
\title{SPI Normalisation Data}
\format{
A data frame with 4 columns and 55 rows
\describe{
\item{Indicator}{Indicator name}
\item{Inverted}{TRUE if the indicator is inverted, FALSE otherwise}
\item{Utopian value}{Official utopian value for each indicator}
\item{Dystopian value}{Official dystopian value for each indicator}
}
}
\source{
https://ec.europa.eu/regional_policy/sources/work/202006_spi_en.pdf
}
\usage{
spi_normalisation_data
}
\description{
A dataframe containing all the normalisation data needed for the EU-SPI (2020)
replication.
}
\keyword{datasets}
