% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi_normalisation_m.R
\name{spi_normalisation_m}
\alias{spi_normalisation_m}
\title{Dataframe Normalisation EU-SPI (2020)}
\usage{
spi_normalisation_m(type, normdata = NULL, rawdata = NULL)
}
\arguments{
\item{type}{Variable that accepts the following values:
\itemize{
\item "regional" to normalise values for the NUTS-2 EU regions;
\item "national" to normalise values for the EU state members.
}}

\item{normdata}{Normalisation data needed according to the EU-SPI (2020) methodology.
By default, it uses the official normalisation data provided by the European
Commission. See details to learn how to use custom normalisation data.}

\item{rawdata}{Data frame of values that need to be normalised. By default,
it uses the official raw indicator data provided by the European Commission.}
}
\description{
Normalise data frames according to the EU-SPI (2020) methodology.
}
\details{
\subsection{Custom Normalisation Data}{

The custom normalisation data needs to be a data frame containing 4 columns, as such:
\itemize{
\item First column: (character) Should have the indicator name, one for each row.
\item Second column: (logical) Should be TRUE if the indicator is inverted (negative relation
to the index) or FALSE otherwise (positive relation to the index).
\item Third column: (numeric) Should have the utopian values for each indicator.
\item Fourth column: (numeric) Should have the dystopian values for each indicator.
}
}

\subsection{Custom Raw Data}{

The function expects the raw data to have one indicator per column and one row per
observation. All columns should be numeric. At the start of the dataframe, however,
the firsts columns can be character for labeling purposes -like the NUTS IDs or the
region names- and they will be preserved after the normalisation process.
}
}
\examples{
#With the official data
spi_normalisation_m(type = "national")

#With custom data
normalisation_data <- data.frame(indicators = c("Indicator 1", "Indicator 2", "Indicator 3"),
                                 inverted = c(TRUE, FALSE, TRUE),
                                 utopian = c(0,100,24),
                                 dystopian = c(100,0,72))
raw_data <- data.frame(regions = c("Region A", "Region B"),
                       indicator1 = sample(0:100,2),
                       indicator2 = sample(0:100,2),
                       indicator3 = sample(24:72,2))
spi_normalisation_m(type = "regional", normdata = normalisation_data, rawdata = raw_data)

}
