% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi_rawdata_down.R
\name{spi_rawdata_down}
\alias{spi_rawdata_down}
\title{Download the EU-SPI (2020) Raw Data}
\usage{
spi_rawdata_down(type)
}
\arguments{
\item{type}{Variable that accepts the values "REGIONAL" to download the regional
data or "NATIONAL to download the national data. It can also take the value
"NORM" to provide the normalisation data from the GitHub of the EUspiR
package, made from the European Commision data.}
}
\value{
A data frame with the chosen data type
}
\description{
This downloads the official raw data used for the creation of the EU-SPI (2020)
as presented in the European Commission page. It returns a data frame with the
values for each indicator and each region at NUTS-2 level or country at NUTS-0.
Additionaly, it returns the official normalisation data.
}
\references{
Source of the indicator data: https://ec.europa.eu/regional_policy/sources/work/spi2020_raw_data.xlsx

Source of the normalisation data: https://ec.europa.eu/regional_policy/sources/work/202006_spi_en.pdf
}
