% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_spi.R
\name{weighted_spi}
\alias{weighted_spi}
\title{Calculation of the weighted components within EU-SPI (2020)}
\usage{
weighted_spi(country, unweighted_values = NULL, nationalValue = NULL)
}
\arguments{
\item{country}{NUTS-0 code of the country to weight. For several countries,
country should be a vector of the NUTS-0 codes to be used. It can also take the value
"all" so that it is applied to each EU country at once.}

\item{unweighted_values}{Data frame with the values to weight of the chosen
component. See details for more information about the format of the data frame.}

\item{nationalValue}{Data frame with the national values to weight for the chosen
component. See details for more information about the format of the data frame.}
}
\description{
Calculate the weighted value of a given component with the population data of
the NUTS-2 regions, according to the EU-SPI (2020) methodology.
}
\details{
The parameters unweighted_values and nationalValue dataframe should have the following
format:
\itemize{
\item First column(s) should be of type character. There is no rules on how many
columns should the data frame have of this type but the function expects at
least one to be named "NUTS_ID" and have the NUTS-2 or NUT-S0 code for each
region on each row.
\item The following columns should be of type numeric. Each column should have one
component to be weighted and each row should represent one region or country.
}
}
\examples{
data <- data.frame("NUTS_ID" = c("AT11", "AT12", "AT13", "AT21",
                                 "AT22", "AT31", "AT32", "AT33", "AT34"),
                   "Component1" = sample(1:100,9),
                   "Component2" = sample(1:100,9),
                   "Component3" = sample(1:100,9))
national_val <- data.frame("NUTS_ID" = "AT",
                           "Component1" = 52,
                           "Component2" = 49,
                           "Component3" = 55)
weighted_spi("AT", data, national_val)

}
