% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi_indicator_sum.R
\name{spi_indicator_sum}
\alias{spi_indicator_sum}
\title{Aggregation of indicators within the EU-SPI (2020)}
\usage{
spi_indicator_sum(indicators, componentName)
}
\arguments{
\item{indicators}{Data frame with the values of each indicator (columns) for each
region (rows)}

\item{componentName}{Name of the component that is to be obtained as the final
result of the aggregation}
}
\description{
Aggregate a set of indicators as seen in the EU-SPI (2020) methodology
}
\examples{
data <- data.frame("Regions" = c("Region A", "Region B", "Region C"),
                   "Indicator 1" = sample(1:100,3),
                   "Indicator 2" = sample(1:100,3),
                   "Indicator 3" = sample(1:100,3))
spi_indicator_sum(data,"Component 1")

}
