% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi_indicator_sum.R
\name{spi_indicator_sum}
\alias{spi_indicator_sum}
\title{Aggregation of indicators within the EU-SPI (2020)}
\usage{
spi_indicator_sum(
  type,
  data = NULL,
  component_name = "all",
  component_index = NULL,
  append = TRUE
)
}
\arguments{
\item{type}{Variable that can take the values "national" when the values to
aggregate are at the NUTS-0 level or "regional" when the values to aggregate are
at the NUTS-2 level.}

\item{data}{Data frame with the values of each indicator (column), for each
region or country (rows). By default (data = NULL) it takes the official
data from the European Commission.}

\item{component_name}{Name of the component(s) to be obtained as the final result
of the aggregation. Alternatively, and by default, setting it to "all" uses the
names of the twelve official components on the EU-SPI (2020).}

\item{component_index}{List of the column indeces of each component within data.
Each item of the list should be a numeric vector stating the indeces of each
component's indicators. By default it will assume that the index follows the
official data when data is not provided or that all the indicators are to be
aggregated to only one component when data is provided.}

\item{append}{Logical argument. By default (TRUE) retains the values of all
indicators alongside the aggregated values of each component. When set to FALSE,
only the aggregated values are kept.}
}
\description{
Aggregate a set of indicators as seen in the EU-SPI (2020) methodology.
}
\examples{
#With the official data
regional_components <- spi_indicator_sum("regional")
national_components <- spi_indicator_sum("national")

#With custom data
data <- data.frame("Regions" = c("Region A", "Region B", "Region C"),
                   "Indicator 1" = sample(1:100,3),
                   "Indicator 2" = sample(1:100,3),
                   "Indicator 3" = sample(1:100,3))
spi_indicator_sum("regional", data,"Component 1")

}
