% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi_rawdata_down.R
\name{spi_rawdata_down}
\alias{spi_rawdata_down}
\title{Download EU-SPI (2020) Official Data}
\usage{
spi_rawdata_down(type)
}
\arguments{
\item{type}{Variable that accepts the following values:
\itemize{
\item "regional", to download the regional raw indicator data;
\item "national", to download the national raw indicator data;
\item "normalisation", to download the normalisation data.
}}
}
\value{
Returns either a data frame with the values for each indicator at the
NUTS-2 level if type is "regional" or at the NUTS-0 level (country level) if
type is "national". If type is "normalisation" it returns a data frame with the
utopian and dystopian values and whether the indicator is inverted for each indicator
of the EU-SPI 2020 instead.
}
\description{
Download the official raw and normalisation data used for the creation of the
EU-SPI (2020) as presented by the European Commision in the offical web page.
}
\references{
Source of the indicator data: https://ec.europa.eu/regional_policy/sources/work/spi2020_raw_data.xlsx

Source of the normalisation data: https://ec.europa.eu/regional_policy/sources/work/202006_spi_en.pdf
}
