% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_spi.R
\name{weighted_spi}
\alias{weighted_spi}
\title{Calculation of the weighted components within EU-SPI (2020)}
\usage{
weighted_spi(
  country,
  unweighted_values = NULL,
  national_value = NULL,
  append = TRUE
)
}
\arguments{
\item{country}{NUTS-0 code of the country to weight. For more than one country,
it should be a string vector of the NUTS-0 codes to be used. It can also take
the value "all" so that it is applied to all EU countries at once.}

\item{unweighted_values}{Data frame with the values to weight. By default it
uses the official EU-SPI (2020) data. See details to learn how to use custom
data.}

\item{national_value}{Data frame with the national values to use. By default it
uses the official EU-SPI (2020) data. See details to learn how to use custom
data.}
}
\description{
Calculate the weighted value of a given component with the population data of
the NUTS-2 regions, according to the EU-SPI (2020) methodology.
}
\details{
\subsection{Custom unweighted values}{
\itemize{
\item First column(s): At least one of the first columns should be called "NUTS_ID",
this column should have the NUTS-2 code for each region to weight.
\item The following columns should be numeric and each one should have one component
to be weighted. Each row should represent one region.
}
}

\subsection{Custom national values}{
\itemize{
\item First column(s): At least one of the first columns should be called "NUTS_ID",
and have the NUTS-0 code for each country.
\item The following columns should be numeric and each one should have the national
values of one component. Each row should represent one country.
}
}
}
\examples{
# With official data
## For one country
weighted_spi("AT") # Weighted values for Austria
## For several countries
weighted_spi(c("BE","ES")) # Weighted values for Belgium and Spain
## For all countries
weighted_spi("all")

# With custom data
data <- data.frame("NUTS_ID" = c("AT11", "AT12", "AT13", "AT21",
                                 "AT22", "AT31", "AT32", "AT33", "AT34"),
                   "Component1" = sample(1:100,9),
                   "Component2" = sample(1:100,9),
                   "Component3" = sample(1:100,9))
national_val <- data.frame("NUTS_ID" = "AT",
                           "Component1" = 52,
                           "Component2" = 49,
                           "Component3" = 55)
weighted_spi("AT", data, national_val)

}
